craigslist/requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coverage==6.2
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.0
django-braces==1.15.0
django-elasticsearch-dsl==7.2.1
django-oauth-toolkit==1.6.1
django-redis==5.2.0
django-rest-elasticsearch==0.4.2
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
elasticsearch==7.16.2
elasticsearch-dsl==7.4.0
factory-boy==3.2.1
Faker==11.1.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
jwcrypto==1.0
msgpack==1.0.3
oauthlib==3.1.1
outcome==1.1.0
packaging==21.3
psycopg2==2.8.6
pycparser==2.21
PyJWT==2.3.0
pynvim==0.4.3
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
redis==4.1.0
requests==2.26.0
requests-oauthlib==1.3.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
sqlparse==0.4.1
text-unidecode==1.3
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wrapt==1.13.3
wsproto==1.0.0
